chore(tool): consolidate instrument package tests into per-source-file pairing
Last package in the test-consolidation pass for #1128. tool/internal/instrument
had ten orphaned test files:

  api_sync_test.go               -> apply_func_test.go (templateAPI is in apply_func.go)
  call_placeholder_test.go       -> call_template_test.go (replacePlaceholder)
  call_template_extra_test.go    -> call_template_test.go (compileExpression,
                                     parseGoExpression, parseGoTypeExpression)
  hookcontext_runtime_test.go    -> trampoline_test.go (simulates the
                                     HookContextImpl trampoline.go generates)
  impl_template_test.go          -> trampoline_test.go (templateImpl)
  trampoline_generic_test.go     -> trampoline_test.go (isTypeParameter,
                                     replaceTypeParamsWithAny)
  trampoline_receiver_test.go    -> trampoline_test.go (extractReceiverTypeParams)
  tool_version_test.go splits two ways:
    TestToolVersionLine, TestMarkedToolVersion, TestEnableNestedToolexec,
    TestToolVersionLineMatchesCachePattern -> toolexec_test.go (all four
    target functions are declared in toolexec.go)
    TestLoadMissingMatchedRules -> new match_test.go (ip.load() is declared
    in match.go, which had no test file at all)
  toolexec_extra_test.go         -> toolexec_test.go

Two orphans are deliberately left as-is:

  toolexec_exec_test.go carries `//go:build !windows` (it shells out to
  writeStub, which can't run on Windows); the maintainer already confirmed
  on #1128 that platform-constrained files should stay their own file
  rather than be folded into their unconstrained sibling.

  logger_delegate_test.go has no build constraint of its own, but its
  natural target, instrument_test.go, carries `//go:build !windows`.
  Merging it in would silently drop its Windows coverage, exactly the
  tradeoff #1128 flagged as needing an explicit answer rather than a
  silent conversion. Left alone pending that call.

This is a pure move for everything actually merged: no test was added,
removed, renamed, or edited. `go test -v ./tool/internal/instrument/...`
produces an identical {name, pass/fail} set before and after (154 entries,
diffed clean), run under WSL so the !windows-gated files are included in
both runs and the comparison is apples-to-apples.
